first four functions moved and removed unnecessory imports
