+++ readthedocs/projects/tasks.py
@@ -39,7 +39,7 @@ from readthedocs.builds.constants import (LATEST,
 from readthedocs.builds.models import Build, Version, APIVersion
 from readthedocs.builds.signals import build_complete
 from readthedocs.builds.syncers import Syncer
-from readthedocs.builds.tasks import fileify
+from readthedocs.builds.tasks import fileify, send_notifications, email_notification
 from readthedocs.cdn.purge import purge
 from readthedocs.core.resolver import resolve_path
 from readthedocs.core.symlink import PublicSymlink, PrivateSymlink
@@ -903,93 +903,6 @@ def _manage_imported_files(version, path, commit):
                 version.project, filename=fname, version_slug=version.slug,
             ) for fname in changed_files]
             purge(cdn_ids[version.project.slug], changed_files)
-
-
-@app.task(queue='web')
-def send_notifications(version_pk, build_pk):
-    version = Version.objects.get(pk=version_pk)
-    build = Build.objects.get(pk=build_pk)
-
-    for hook in version.project.webhook_notifications.all():
-        webhook_notification(version, build, hook.url)
-    for email in version.project.emailhook_notifications.all().values_list('email', flat=True):
-        email_notification(version, build, email)
-
-
-def email_notification(version, build, email):
-    """
-    Send email notifications for build failure.
-
-    :param version: :py:class:`Version` instance that failed
-    :param build: :py:class:`Build` instance that failed
-    :param email: Email recipient address
-    """
-    log.debug(LOG_TEMPLATE.format(project=version.project.slug, version=version.slug,
-                                  msg='sending email to: %s' % email))
-
-    # We send only what we need from the Django model objects here to avoid
-    # serialization problems in the ``readthedocs.core.tasks.send_email_task``
-    context = {
-        'version': {
-            'verbose_name': version.verbose_name,
-        },
-        'project': {
-            'name': version.project.name,
-        },
-        'build': {
-            'pk': build.pk,
-            'error': build.error,
-        },
-        'build_url': 'https://{0}{1}'.format(
-            getattr(settings, 'PRODUCTION_DOMAIN', 'readthedocs.org'),
-            build.get_absolute_url(),
-        ),
-        'unsub_url': 'https://{0}{1}'.format(
-            getattr(settings, 'PRODUCTION_DOMAIN', 'readthedocs.org'),
-            reverse('projects_notifications', args=[version.project.slug]),
-        ),
-    }
-
-    if build.commit:
-        title = _('Failed: {project[name]} ({commit})').format(commit=build.commit[:8], **context)
-    else:
-        title = _('Failed: {project[name]} ({version[verbose_name]})').format(**context)
-
-    send_email(
-        email,
-        title,
-        template='projects/email/build_failed.txt',
-        template_html='projects/email/build_failed.html',
-        context=context,
-    )
-
-
-def webhook_notification(version, build, hook_url):
-    """
-    Send webhook notification for project webhook.
-
-    :param version: Version instance to send hook for
-    :param build: Build instance that failed
-    :param hook_url: Hook URL to send to
-    """
-    project = version.project
-
-    data = json.dumps({
-        'name': project.name,
-        'slug': project.slug,
-        'build': {
-            'id': build.id,
-            'success': build.success,
-            'date': build.date.strftime('%Y-%m-%d %H:%M:%S'),
-        }
-    })
-    log.debug(LOG_TEMPLATE
-              .format(project=project.slug, version='',
-                      msg='sending notification to: %s' % hook_url))
-    try:
-        requests.post(hook_url, data=data)
-    except Exception:
-        log.exception('Failed to POST on webhook url: url=%s', hook_url)
 
 
 @app.task(queue='web')
